Extract account-info crate (#2429)
* extract account-info crate

* make bincode optional in account-info crate

* update lock file

* update doc examples

* remove solana-program from doc examples

* remove solana-program from dev deps

* fmt

* move MAX_PERMITTED_DATA_INCREASE to account-info crate
