+++ Cargo.toml
@@ -101,6 +101,7 @@ members = [
     "runtime-transaction",
     "sanitize",
     "sdk",
+    "sdk/account-info",
     "sdk/atomic-u64",
     "sdk/cargo-build-sbf",
     "sdk/cargo-test-sbf",
@@ -359,6 +360,7 @@ smpl_jwt = "0.7.1"
 socket2 = "0.5.7"
 soketto = "0.7"
 solana-account-decoder = { path = "account-decoder", version = "=2.1.0" }
+solana-account-info = { path = "sdk/account-info", version = "=2.1.0" }
 solana-accounts-db = { path = "accounts-db", version = "=2.1.0" }
 solana-address-lookup-table-program = { path = "programs/address-lookup-table", version = "=2.1.0" }
 solana-atomic-u64 = { path = "sdk/atomic-u64", version = "=2.1.0" }
