+++ sdk/program/src/entrypoint.rs
@@ -5,7 +5,6 @@
 //! [`bpf_loader`]: crate::bpf_loader
 
 extern crate alloc;
-pub use solana_program_error::ProgramResult;
 use {
     crate::{account_info::AccountInfo, pubkey::Pubkey},
     alloc::vec::Vec,
@@ -18,6 +17,7 @@ use {
         slice::{from_raw_parts, from_raw_parts_mut},
     },
 };
+pub use {solana_account_info::MAX_PERMITTED_DATA_INCREASE, solana_program_error::ProgramResult};
 
 /// User implemented function to process an instruction
 ///
@@ -309,9 +309,6 @@ unsafe impl std::alloc::GlobalAlloc for BumpAllocator {
         // I'm a bump allocator, I don't free
     }
 }
-
-/// Maximum number of bytes a program may add to an account during a single realloc
-pub const MAX_PERMITTED_DATA_INCREASE: usize = 1_024 * 10;
 
 /// `assert_eq(std::mem::align_of::<u128>(), 8)` is true for BPF but not for some host machines
 pub const BPF_ALIGN_OF_U128: usize = 8;
